virtio-gpu: carry a range descriptor's horizontal rate past 255 kHz
The EDID range descriptor stores horizontal rate in a single byte of kHz, and a
4K panel at 120 Hz is 265 — so the bound had to be clamped, which understates it.
That is the harmful direction: the guest checks candidate modes against the range
and prunes anything above it, so a clamped bound prunes the very mode we
advertise alongside it.

EDID 1.4 has a "+255" offset flag per bound in byte 4 for exactly this. Emit them,
and widen the horizontal fields to u16 so a caller can express the value at all.

The vertical rates keep their plain bytes: nothing we advertise approaches 255 Hz,
and Linux only applies the vertical offsets anyway (drm_edid.c, get_monitor_range).

+++ src/devices/src/display/types.rs
@@ -66,8 +66,10 @@ pub struct EdidIdentity {
 pub struct RefreshRange {
     pub min_vertical_hz: u8,
     pub max_vertical_hz: u8,
-    pub min_horizontal_khz: u8,
-    pub max_horizontal_khz: u8,
+    /// Horizontal rates in kHz. Wider than the descriptor's byte on purpose: EDID 1.4
+    /// carries a +255 kHz offset flag per bound, and a 4K panel at 120 Hz needs it.
+    pub min_horizontal_khz: u16,
+    pub max_horizontal_khz: u16,
     /// Maximum pixel clock in MHz; stored in the descriptor rounded up to a 10 MHz step.
     pub max_pixel_clock_mhz: u32,
 }
